fix and simplify config
(cherry picked from commit 9c78a524e8df2b60b9b528dcf55e4e8d1beb253f)

+++ src/main/java/de/dertoaster/crossbowverhaul/enchantment/MultishotEnchantment.java
@@ -10,7 +10,7 @@ public class MultishotEnchantment extends net.minecraft.enchantment.MultishotEnc
 	
 	@Override
 	public int getMaxLevel() {
-		return CrossbowverhaulConfigHolder.coModMultishot.get();
+		return CrossbowverhaulConfigHolder.coModMultishot.get().intValue();
 	}
 
 }
